pyproject.toml
[tool.poetry]
name = "nome-do-projeto"
version = "0.1.0"
description = ""
authors = ["EvaReboucas30 <140523579+EvaReboucas30@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "nome_do_projeto", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
djangorestframework = "^3.16.0"
drf-yasg = "^1.21.10"
psycopg2-binary = "^2.9.10"
django-jazzmin = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
